fix(eraser): Prevent erasing SVG elements
This commit introduces a check within the eraser tool to prevent the accidental erasure of SVG elements. If an attempt is made to erase an SVG element, a notification will be displayed to the user, and the SVG element will be ignored.

This change also updates the localization files for both English and Tamil to include the new error message.