feat(DrawingModal): Allow base glyphs to be marks in composites
Previously, composite characters with positioning rules only considered base glyphs as 'base' class characters. This change updates the logic to allow a 'mark' class character to also function as the base glyph in a two-character composite, provided the second character is explicitly a 'mark'. This expands the flexibility for creating composite characters with complex positioning.